requirements.txt
# Core dependencies
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0

# Statistical analysis
statsmodels>=0.14.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Natural Language Processing
bert-score>=0.3.13
torch>=2.0.0
transformers>=4.21.0
tokenizers>=0.13.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.64.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional utilities
requests>=2.28.0
argparse  # Built-in, but listed for clarity 